新增后台文章发布页面

+++ src/main/java/com/neilren/neilren4j/modules/article/dao/ArticleDao.java
@@ -14,7 +14,7 @@ import java.util.Map;
 public interface ArticleDao extends CrudDao<Article> {
     int deleteByPrimaryKey(Long id);
 
-    int insert(ArticleWithBLOBs record);
+    Long insert(ArticleWithBLOBs record);
 
     int insertSelective(ArticleWithBLOBs record);
